consistent integer size

+++ src/Hash.h
@@ -1,8 +1,8 @@
 #ifndef HASH_H
 #define HASH_H
-
+#include <stdint.h>
 #define HASH_LENGTH 243
 #define STATE_LENGTH 3 * HASH_LENGTH
-typedef long long trit_t;
+typedef int64_t trit_t;
 
 #endif
